Add factual reporting to icon
The background of the icon is determined by adding together the bias
score and the factual score.

Green is used for the least biased/most trustworthy sources (~16%).
Pink/Purple is used for the semi-biased/mostly trustworthy sources (~25%).
Orange is used for heavily biased/untrustworthy sources (~59%).

Orange is bad; Green is good; Pink is okay